stored-procedures-implementation: implement functions closeorder, createrequest, closerequest & need to fix getorderdetailsfortable function implementation

+++ demo/src/main/java/com/monsatorm/demo/model/projections/OrderDetailDtoP.java
@@ -1,11 +1,13 @@
 package com.monsatorm.demo.model.projections;
 
+import java.math.BigDecimal;
+
 public interface OrderDetailDtoP {
     String getProductName();
 
     Integer getQuantity();
 
-    Integer getPrice();
+    BigDecimal getPrice();
 
-    Integer getTotalPrice();
+    BigDecimal getTotalPrice();
 }
